requirements.txt
streamlit
pillow
azure-ai-formrecognizer
azure-core